pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "vision-parse"
dynamic = ["version"]
description = "Parse PDF documents into markdown formatted content using Vision LLMs"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Arun Brahma", email = "mithubrahma@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "jinja2>=3.0.0",
    "pydantic>=2.0.0",
    "pymupdf>=1.22.0",
    "tenacity>=9.0.0",
    "tqdm>=4.65.0",
]

[project.urls]
Homepage = "https://github.com/iamarunbrahma/vision-parse"
Repository = "https://github.com/iamarunbrahma/vision-parse.git"

[project.optional-dependencies]
dev = [
    "black>=24.4.1",
    "pytest>=8.3.4",
    "ruff>=0.8.3",
]
openai = [
    "openai==1.58.0",
]
ollama = [
    "ollama>=0.4.4",
]
gemini = [
    "google-generativeai==0.8.3",
]
all = [
    "google-generativeai==0.8.3",
    "ollama>=0.4.4",
    "openai==1.58.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vision_parse"]

[tool.hatch.build.targets.sdist]
include = [
    "/src/vision_parse",
    "/tests",
]

[tool.hatch.build.targets.wheel.force-include]
"src/vision_parse/md_prompt.j2" = "vision_parse/md_prompt.j2"
"src/vision_parse/img_analysis.prompt" = "vision_parse/img_analysis.prompt"

[tool.hatch.version]
source = "vcs"